requirements.txt
streamlit
sentence-transformers
faiss-cpu # or faiss-gpu
Pillow
requests
tesserocr
